requirements.txt
numpy
opencv-python
pillow
flask
tensorflow